services/auth-proxy/requirements.txt
# Auth Proxy Service Dependencies

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.25.0

# JWT handling
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Database (shared with main project)
# These will be installed from parent requirements.txt
# sqlalchemy[asyncio]>=2.0.23
# asyncpg>=0.29.0
# alembic>=1.12.1

# Redis (for future session management)
# redis>=5.0.1

# Logging and monitoring
python-json-logger>=2.0.7
